+++ src/constructor.js
@@ -1,7 +1,9 @@
-export default class Object {
+class Object {
   constructor(description, completed, index) {
     this.description = description;
     this.completed = completed;
     this.index = index;
   }
 }
+
+export default Object;
